+++ R/save_tt.R
@@ -86,8 +86,8 @@ save_tt <- function(x, output, overwrite = FALSE) {
     assert_dependency("tinytex")
     # \documentclass{standalone} does not support \begin{table}
     tmp <- strsplit(x@table_string, "\\n")[[1]]
-    tmp <- tmp[!grepl("\\begin{table}", tmp, fixed = TRUE)]
-    tmp <- tmp[!grepl("\\end{table}", tmp, fixed = TRUE)]
+    # tmp <- tmp[!grepl("\\begin{table}", tmp, fixed = TRUE)]
+    # tmp <- tmp[!grepl("\\end{table}", tmp, fixed = TRUE)]
     tmp <- paste(tmp, collapse = "\n")
     tmp <- sprintf(latex_standalone, tmp)
     # tinytex is fiddly with file paths, so we need to hack 
@@ -124,6 +124,8 @@ latex_standalone <- "
 \\newcommand{\\tinytableTabularrayStrikeout}[1]{\\sout{#1}}
 \\NewTableCommand{\\tinytableDefineColor}[3]{\\definecolor{#1}{#2}{#3}}
 \\begin{document}
+\\minipage{\\textwidth}
 %s
+\\endminipage
 \\end{document}
 "